kmer_indel.pl
This perl script replaces both kmer_del.pl and kmer_ins.pl.  In this
step one generates an enrichment value using the total number of kmers
that span a gap, relative to the total number of kmers in a given
reference sequence (fast format).  This differs from the previous
version since the program no longer utilizes a random null. Assumptions
of the program:  (a) the provided reference file contains a single fast
sequence in the reference file and (b) the length of k is fixed at 5bp.
 Two output files are generated from this step:  cmpKmer.del.txt and
cmpKmer.ins.txt